:recycle: Reformat RNN class structure
